+++ frontend/apps/mobile/babel.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-    presets: ["module:metro-react-native-babel-preset"],
-    plugins: ["react-native-paper/babel", "react-native-reanimated/plugin"],
-};