+++ frontend/apps/mobile/src/@types/modules.d.ts
@@ -1,2 +1,0 @@
-declare module "react-native-vector-icons/MaterialCommunityIcons";
-declare module "react-native-vector-icons/MaterialIcons";