+++ frontend/apps/mobile/src/main.tsx
@@ -1,5 +1,0 @@
-import { AppRegistry } from "react-native";
-import { AppWrapper } from "./AppWrapper";
-import "./i18n";
-
-AppRegistry.registerComponent("Abrechnung", () => AppWrapper);