+++ frontend/apps/mobile/src/store/index.ts
@@ -1,3 +1,0 @@
-export * from "./store";
-export * from "./settingsSlice";
-export * from "./uiSlice";